refactor: split repo_handler.py into validation and clone modules
